requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-WTF
python-dotenv